[010/c] fixes solution

+++ src/010-summation-of-primes/c/solve.c
@@ -1,10 +1,10 @@
 #include "solve.h"
 
-uint32_t solve(uint32_t n) {
+uint64_t solve(uint64_t n) {
     prime_t p = prime_new();
     size_t i = 1;
-    uint32_t sum = 0;
-    uint32_t prime = 2;
+    uint64_t sum = 0;
+    uint64_t prime = 2;
 
     while(prime < n) {
         sum += prime;
